+++ changes.md
@@ -1,15 +1,36 @@
 #### Starting from 04/02/2023
 
-### 05/30/23
-new file:   CODEFORCES/1200/1738B.cpp
-new file:   CODEFORCES/1200/1742E.cpp
+### 06/04/23
+new file:   CODEFORCES/1000/1838B.cpp
+new file:   CODEFORCES/1000/1839A.cpp
+new file:   CODEFORCES/1100/1838C.cpp
 
+### 06/03/23
+new file:   CODEFORCES/1100/1839B.cpp
+new file:   CODEFORCES/1200/1720C.cpp
+new file:   CODEFORCES/1200/1839C.cpp
+
+### 06/02/23
+new file:   CODEFORCES/1300/1717C.cpp
+new file:   CODEFORCES/1300/1726C.cpp
+new file:   CODEFORCES/900/1838A.cpp
+new file:   USACO/Guide/1-3-Dual-Palindrome.cpp
+
+### 06/01/23
+- new file:   CODEFORCES/1200/1729D.cpp
+- new file:   CODEFORCES/1200/1730C.cpp
+- new file:   CODEFORCES/1200/1734C.cpp
+- new file:   CODEFORCES/1300/1736C1.cpp
+
+### 05/30/23
+- new file:   CODEFORCES/1200/1738B.cpp
+- new file:   CODEFORCES/1200/1742E.cpp
 
 ### 05/29/23
-new file:   CODEFORCES/1200/1744D.cpp
-new file:   CODEFORCES/1300/1741D.cpp
-new file:   CODEFORCES/1300/1746C.cpp
-new file:   USACO/Guide/1-3-Palindrome-square
+- new file:   CODEFORCES/1200/1744D.cpp
+- new file:   CODEFORCES/1300/1741D.cpp
+- new file:   CODEFORCES/1300/1746C.cpp
+- new file:   USACO/Guide/1-3-Palindrome-square
 
 ### 05/28/2023
 - new file:   USACO/Guide/1-3-Name-that-number.cpp